pyproject.toml
[tool.poetry]
name = "district"
version = "0.1"
description = "District_Web_App"
authors = ["proxod3-first"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aioredis = "1.3.1"
asgiref = "3.4.1"
async-timeout = "4.0.2"
attrs = "23.1.0"
autobahn = "23.6.2"
automat = "22.10.0"
autopep8 = "1.5.7"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2023.11.17"
cffi = "1.16.0"
channels = "3.0.4"
channels-redis = "3.4.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
constantly = "23.10.4"
coverage = "7.3.2"
crashtest = "0.4.1"
cryptography = "41.0.7"
daphne = "3.0.2"
distlib = "0.3.7"
django = "3.2.7"
django-cors-headers = "3.8.0"
django-redis = "5.4.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
exceptiongroup = "1.2.0"
fastjsonschema = "2.19.0"
filelock = "3.13.1"
hiredis = "2.1.0"
hyperlink = "21.0.0"
idna = "3.6"
importlib-metadata = "7.0.0"
incremental = "22.10.0"
iniconfig = "2.0.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
keyring = "24.3.0"
more-itertools = "10.1.0"
msgpack = "1.0.7"
packaging = "23.2"
pexpect = "4.9.0"
pillow = "10.1.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
pluggy = "1.3.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycodestyle = "2.7.0"
pycparser = "2.21"
pyjwt = "2.8.0"
pyopenssl = "23.3.0"
pyproject-hooks = "1.0.0"
pytest = "7.4.3"
python-decouple = "3.8"
pytz = "2021.1"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.5.2"
redis = "5.0.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
service-identity = "23.1.0"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.4.2"
toml = "0.10.2"
tomli = "2.0.1"
tomlkit = "0.12.3"
trove-classifiers = "2023.11.29"
twisted = "23.10.0"
txaio = "23.1.1"
typing-extensions = "4.8.0"
urllib3 = "2.1.0"
virtualenv = "20.25.0"
zipp = "3.17.0"
zope-interface = "6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
